+++ dashboard/src/components/partials/CheckDetailsContainer/CheckDetailsConfiguration.js
@@ -21,6 +21,7 @@ import CodeHighlight from "/components/CodeHighlight/CodeHighlight";
 import SilencedIcon from "/icons/Silence";
 import Typography from "@material-ui/core/Typography";
 import Tooltip from "@material-ui/core/Tooltip";
+import LabelsAnnotationsCell from "/components/partials/LabelsAnnotationsCell";
 
 class CheckDetailsConfiguration extends React.PureComponent {
   static propTypes = {
@@ -65,8 +66,9 @@ class CheckDetailsConfiguration extends React.PureComponent {
 
         proxyEntityName
         proxyRequests {
-          entityAttributes
           splay
+          splayCoverage
+          entityAttributes
         }
 
         assets: runtimeAssets {
@@ -75,8 +77,12 @@ class CheckDetailsConfiguration extends React.PureComponent {
         }
 
         envVars
-        extendedAttributes
+
+        metadata {
+          ...LabelsAnnotationsCell_objectmeta
+        }
       }
+      ${LabelsAnnotationsCell.fragments.objectmeta}
     `,
   };
 
@@ -236,11 +242,27 @@ class CheckDetailsConfiguration extends React.PureComponent {
                   </DictionaryValue>
                 </DictionaryEntry>
 
-                <DictionaryEntry>
+                <DictionaryEntry fullWidth={!!check.proxyRequests}>
                   <DictionaryKey>Proxy Requests</DictionaryKey>
-                  <DictionaryValue>
+                  <DictionaryValue explicitRightMargin>
                     <Maybe value={check.proxyRequests} fallback="None">
-                      {val => JSON.stringify(val)}
+                      {val => (
+                        <CodeBlock>
+                          <CodeHighlight
+                            language="json"
+                            code={JSON.stringify(
+                              {
+                                entity_attributes: val.entityAttributes,
+                                splay: val.splay,
+                                splay_coverage: val.splayCoverage,
+                              },
+                              null,
+                              "  ",
+                            )}
+                            component="code"
+                          />
+                        </CodeBlock>
+                      )}
                     </Maybe>
                   </DictionaryValue>
                 </DictionaryEntry>
@@ -264,9 +286,9 @@ class CheckDetailsConfiguration extends React.PureComponent {
                   </DictionaryValue>
                 </DictionaryEntry>
 
-                <DictionaryEntry>
+                <DictionaryEntry fullWidth={check.envVars.length > 0}>
                   <DictionaryKey>ENV Vars</DictionaryKey>
-                  <DictionaryValue scrollableContent>
+                  <DictionaryValue>
                     {check.envVars.length > 0 ? (
                       <CodeBlock>
                         <CodeHighlight
@@ -349,24 +371,8 @@ class CheckDetailsConfiguration extends React.PureComponent {
           </Grid>
         </CardContent>
 
-        {Object.keys(check.extendedAttributes).length > 0 && (
-          <React.Fragment>
-            <Divider />
-            <CodeBlock>
-              <CardContent>
-                <CodeHighlight
-                  language="json"
-                  code={`# Extra\n\n${JSON.stringify(
-                    check.extendedAttributes,
-                    null,
-                    "\t",
-                  )}`}
-                  component="code"
-                />
-              </CardContent>
-            </CodeBlock>
-          </React.Fragment>
-        )}
+        <Divider />
+        <LabelsAnnotationsCell check={check} />
       </Card>
     );
   }