+++ dashboard/src/components/views/EnvironmentView/EntitiesContent.js
@@ -11,9 +11,7 @@ import Query from "/components/util/Query";
 import { withQueryParams } from "/components/QueryParams";
 import WithWidth from "/components/WithWidth";
 
-// duration used when polling is enabled; set fairly high until we understand
-// the impact.
-const pollInterval = 2500; // 2.5s
+import { pollingDuration } from "/constants/polling";
 
 class EntitiesContent extends React.PureComponent {
   static propTypes = {
@@ -98,7 +96,7 @@ class EntitiesContent extends React.PureComponent {
       <Query
         query={EntitiesContent.query}
         fetchPolicy="cache-and-network"
-        pollInterval={pollInterval}
+        pollInterval={pollingDuration.short}
         variables={variables}
       >
         {this.renderContent}