+++ dashboard/src/components/views/EnvironmentView/EventsContent.js
@@ -12,12 +12,10 @@ import { withQueryParams } from "/components/QueryParams";
 import ToastConnector from "/components/relocation/ToastConnector";
 import WithWidth from "/components/WithWidth";
 
+import { pollingDuration } from "/constants/polling";
+
 // If none given default expression is used.
 const defaultExpression = "has_check";
-
-// duration used when polling is enabled; set fairly high until we understand
-// the impact.
-const pollInterval = 5000; // 5s
 
 class EventsContent extends React.Component {
   static propTypes = {
@@ -112,7 +110,7 @@ class EventsContent extends React.Component {
       <Query
         query={EventsContent.query}
         fetchPolicy="cache-and-network"
-        pollInterval={pollInterval}
+        pollInterval={pollingDuration.short}
         variables={variables}
       >
         {this.renderContent}